Number Parsing Fix
The Ship class will now be able to properly read Orders that it tries to
parse numbers from.

+++ entities/ships/Ship.java
@@ -186,7 +186,7 @@ public class Ship extends Entity implements ControlSystem{
         
         if(order.substring(0,5).equals("(ACC)")){
             //This is an acceleration order
-            setAcceleration(Integer.parseInt(order.substring(5)));
+            setAcceleration(Double.parseDouble(order.substring(5)));
             if(!orders.get(0).getStatus()){
                 orders.remove(0);
             }
